Separate all button logic and all css into corresponding files.
