+++ client/src/App.css
@@ -44,188 +44,6 @@
     width: 300px;
 }*/
 
-.screen {
-    width: 100vw;
-    height: 100%;
-    border: thick solid orange;
-    box-sizing: border-box;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.game {
-    height: 0;
-    overflow: hidden;
-    padding-top: 35%;
-    width: 80%;
-    position: relative;
-    border-style: solid;
-    border-width: 5px;
-    border-color: green;
-    font-size: 3vw;
-}
-
-.game-inside {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%; /* uaw vh for scale*/
-    border-style: solid;
-    border-width: 2px;
-    border-color: indigo;
-    display: flex;
-    justify-content: space-around;
-}
-
-.board {
-    border: medium solid #c2c2cc;
-    height: 100%;
-    width: calc((35/80) *100%);
-    display: grid;
-    grid-template: 1fr 1fr 1fr / 1fr 1fr 1fr;
-    gap: 0.5%;
-    background-color: #c2c2cc;
-    color: black; /*royalblue;*/
-}
-
-.group {
-    border: thin solid #c2c2cc;
-    background-color: #faf7ff;
-    display: grid;
-    grid-template: 1fr 1fr 1fr / 1fr 1fr 1fr;
-    justify-items: stretch;
-    font-size: 80%;
-    font-weight: 900;
-}
-
-.square {
-    border-style: solid;
-    border-width: 1px;
-    border-color: #c2c2cc;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-weight: 400;
-    color: royalblue;
-    line-height: 0;
-    letter-spacing: 0;
-}
-
-.square-with-marks {
-    border-style: solid;
-    border-width: 1px;
-    border-color: #c2c2cc;
-    display: grid;
-    grid-template: 1fr 1fr 1fr / 1fr 1fr 1fr;
-    grid-template-areas:
-        "1 2 3"
-        "4 5 6"
-        "7 8 9"
-    ;
-    justify-items: stretch;
-    letter-spacing: 0px;
-}
-
-.corner-mark {
-    font-size: 1vw;
-    font-weight: lighter;
-    line-height: 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    color: royalblue;
-}
-
-
-.button-box {
-    border: medium solid darkseagreen;
-    height: 100%;
-    width: calc((35/80) *100%);
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: center;
-}
-
-.button-container {
-    border: medium solid burlywood;
-    height: 55%;
-    width: 77%;
-    display: grid;
-    grid-template: 1fr 1fr 1fr 0.8fr / 1.4fr 1fr 1fr 1fr;
-    grid-template-areas:
-            "controls numpad  numpad  numpad"
-            "controls numpad  numpad  numpad"
-            "controls numpad  numpad  numpad"
-            "footer footer footer footer";
-    font-size: 70%;
-    gap: 1%;
-}
-
-.Controls {
-    border: medium solid orange;
-    grid-area: controls;
-    display: grid;
-    grid-template: 1fr 1fr 1fr 1fr / 1fr;
-    row-gap: 4%;
-    /*display: flex;
-    flex-direction: column;
-    flex-wrap: wrap;
-    justify-content: space-evenly;
-    align-items: center;*/
-}
-
-/*.button-col {
-    height: 100%;
-    width: 100%;
-    border-style: solid;
-    border-width: 1px;
-    border-color: red;
-}*/
-
-/*.button {
-    height: 0;
-    padding: 0 0;
-    font-size: 100%;
-    line-height: 0;
-}*/
-
-.control-button {
-    max-width: 100%;
-    min-width: 100%;
-    max-height: 100%;
-    min-height: 100%;
-    line-height: 0;
-    font-size: 10em;
-    background-color: purple;
-}
-
-.Numpad {
-    border: medium solid dodgerblue;
-    grid-area: numpad;
-    display: grid;
-    grid-template: 1fr 1fr 1fr 1fr / 1fr 1fr 1fr;
-    grid-template-areas:
-            "num  num  num"
-            "num  num  num"
-            "num  num  num"
-            "del  del  del";
-    gap: 4%;
-}
-
-.number-button {
-    max-width: 100%;
-    min-width: 100%;
-    max-height: 100%;
-    min-height: 100%;
-}
-
-.delete-button {
-    grid-area: del;
-}
-
 .Footer {
     height: 90%;
     border: medium solid forestgreen;
@@ -252,8 +70,6 @@
     justify-content: flex-start;
 }
 
-
-
 .button-num-pad {
     width: 70%;
     border-style: solid;
@@ -279,17 +95,3 @@
     height: 100%;
     width: 90%;
 }
-
-.highlight {
-    background-color: rgba(250, 241, 170, 0.5);
-}
-
-.contradict {
-    background-color: red;
-}
-
-.permanent {
-    font-weight: bold;
-    color: black;
-}
-
