+++ client/src/components/ControlButtons.css
@@ -1,0 +1,36 @@
+.Controls {
+    border: medium solid orange;
+    grid-area: controls;
+    display: grid;
+    grid-template: 1fr 1fr 1fr 1fr / 1fr;
+    row-gap: 4%;
+    /*display: flex;
+    flex-direction: column;
+    flex-wrap: wrap;
+    justify-content: space-evenly;
+    align-items: center;*/
+}
+
+/*.button-col {
+    height: 100%;
+    width: 100%;
+    border-style: solid;
+    border-width: 1px;
+    border-color: red;
+}*/
+
+/*.button {
+    height: 0;
+    padding: 0 0;
+    font-size: 100%;
+    line-height: 0;
+}*/
+
+.control-button {
+    max-width: 100%;
+    min-width: 100%;
+    max-height: 100%;
+    min-height: 100%;
+    line-height: 0;
+    font-size: 10em;
+}
