+++ client/src/components/Game.css
@@ -1,0 +1,81 @@
+.screen {
+    width: 100vw;
+    height: 100%;
+    border: thick solid orange;
+    box-sizing: border-box;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.game {
+    height: 0;
+    overflow: hidden;
+    padding-top: 35%;
+    width: 80%;
+    position: relative;
+    border-style: solid;
+    border-width: 5px;
+    border-color: green;
+    font-size: 3vw;
+}
+
+.game-inside {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%; /* uaw vh for scale*/
+    border-style: solid;
+    border-width: 2px;
+    border-color: indigo;
+    display: flex;
+    justify-content: space-around;
+}
+
+.board {
+    border: medium solid #c2c2cc;
+    height: 100%;
+    width: calc((35/80) *100%);
+    display: grid;
+    grid-template: 1fr 1fr 1fr / 1fr 1fr 1fr;
+    gap: 0.5%;
+    background-color: #c2c2cc;
+    color: black; /*royalblue;*/
+}
+
+.group {
+    border: thin solid #c2c2cc;
+    background-color: #faf7ff;
+    display: grid;
+    grid-template: 1fr 1fr 1fr / 1fr 1fr 1fr;
+    justify-items: stretch;
+    font-size: 80%;
+    font-weight: 900;
+}
+
+
+.button-box {
+    border: medium solid darkseagreen;
+    height: 100%;
+    width: calc((35/80) *100%);
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    align-items: center;
+}
+
+.button-container {
+    border: medium solid burlywood;
+    height: 55%;
+    width: 77%;
+    display: grid;
+    grid-template: 1fr 1fr 1fr 0.8fr / 1.4fr 1fr 1fr 1fr;
+    grid-template-areas:
+            "controls numpad  numpad  numpad"
+            "controls numpad  numpad  numpad"
+            "controls numpad  numpad  numpad"
+            "footer footer footer footer";
+    font-size: 70%;
+    gap: 1%;
+}
