+++ client/src/components/Numpad.css
@@ -1,0 +1,25 @@
+.Numpad {
+    border: medium solid dodgerblue;
+    grid-area: numpad;
+    display: grid;
+    grid-template: 1fr 1fr 1fr 1fr / 1fr 1fr 1fr;
+    grid-template-areas:
+            "num  num  num"
+            "num  num  num"
+            "num  num  num"
+            "del  del  del";
+    gap: 4%;
+}
+
+.number-button {
+    max-width: 100%;
+    min-width: 100%;
+    max-height: 100%;
+    min-height: 100%;
+}
+
+.delete-button {
+    grid-area: del;
+    background-color: gray;
+    border: gray;
+}
